PLEASE READ THIS BEFORE CONTINUING WORK
It seems we have been having trouble navigating between scenes, due to this I have created a script called:

SceneFlow.cs


This is to help prevent builds breaking. In order to call a scene, instead of using SceneManager.LoadScene, use:

SceneFlow.RunScene(SceneList.MainMenu);


If you need to add a scene to the game, add it to the enum, and then to the corresponding index on the scenes list. both of these are under the "SceneFlow.cs" file. 

I have changed appropriate functions so far, it should not affect any test scenes or your current build. If this is wrong, let me know ASAP